+++ src/sys/syscall/mod.rs
@@ -1,6 +1,7 @@
 pub mod number;
 pub mod service;
 
+use crate::sys;
 use crate::sys::fs::FileStat;
 
 /*
@@ -20,14 +21,14 @@ pub fn dispatcher(n: usize, arg1: usize, arg2: usize, arg3: usize) -> usize {
             service::realtime().to_bits() as usize
         }
         number::STAT => {
-            let ptr = arg1 as *mut u8;
+            let ptr = sys::process::ptr_from_addr(arg1 as u64);
             let len = arg2;
             let path = unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(ptr, len)) };
             let stat = unsafe { &mut *(arg3 as *mut FileStat) };
             service::stat(path, stat) as usize
         }
         number::OPEN => {
-            let ptr = arg1 as *mut u8;
+            let ptr = sys::process::ptr_from_addr(arg1 as u64);
             let len = arg2;
             let flags = arg3;
             let path = unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(ptr, len)) };
@@ -35,14 +36,14 @@ pub fn dispatcher(n: usize, arg1: usize, arg2: usize, arg3: usize) -> usize {
         }
         number::READ => {
             let handle = arg1;
-            let ptr = arg2 as *mut u8;
+            let ptr = sys::process::ptr_from_addr(arg2 as u64);
             let len = arg3;
             let mut buf = unsafe { core::slice::from_raw_parts_mut(ptr, len) };
             service::read(handle, &mut buf) as usize
         }
         number::WRITE => {
             let handle = arg1;
-            let ptr = arg2 as *mut u8;
+            let ptr = sys::process::ptr_from_addr(arg2 as u64);
             let len = arg3;
             let mut buf = unsafe { core::slice::from_raw_parts_mut(ptr, len) };
             service::write(handle, &mut buf) as usize